Move `UserCardTooltip` to own file
