+++ osu.Game/Users/Drawables/ClickableAvatar.cs
@@ -1,9 +1,7 @@
 ﻿// Copyright (c) ppy Pty Ltd <contact@ppy.sh>. Licensed under the MIT Licence.
 // See the LICENCE file in the repository root for full licence text.
 
-using System.Linq;
 using osu.Framework.Allocation;
-using osu.Framework.Graphics;
 using osu.Framework.Graphics.Containers;
 using osu.Framework.Graphics.Cursor;
 using osu.Framework.Input.Events;
@@ -62,45 +60,6 @@ namespace osu.Game.Users.Drawables
                 return false;
 
             return base.OnClick(e);
-        }
-
-        public partial class UserCardTooltip : VisibilityContainer, ITooltip<APIUser?>
-        {
-            public UserCardTooltip()
-            {
-                AutoSizeAxes = Axes.Both;
-            }
-
-            protected override void PopIn() => this.FadeIn(150, Easing.OutQuint);
-            protected override void PopOut() => this.Delay(150).FadeOut(500, Easing.OutQuint);
-
-            public void Move(Vector2 pos) => Position = pos;
-
-            private APIUser? user;
-
-            public void SetContent(APIUser? content)
-            {
-                if (content == user && Children.Any())
-                    return;
-
-                user = content;
-
-                if (user != null)
-                {
-                    LoadComponentAsync(new UserGridPanel(user)
-                    {
-                        Width = 300,
-                    }, panel => Child = panel);
-                }
-                else
-                {
-                    var tooltip = new OsuTooltipContainer.OsuTooltip();
-                    tooltip.SetContent(ContextMenuStrings.ViewProfile);
-                    tooltip.Show();
-
-                    Child = tooltip;
-                }
-            }
         }
 
         public partial class NoCardTooltip : VisibilityContainer, ITooltip<APIUser?>
